requirements.txt
djangorestframework==3.15.2
asgiref==3.8.1
sqlparse==0.5.0
Django==5.0.6
gunicorn==20.1.0
python-dotenv==1.0.1
